fix: use lodable to avoid media player ssr
