+++ package.json
@@ -36,7 +36,7 @@
   },
   "dependencies": {
     "@colliercz/gatsby-transformer-gitinfo": "github:CollierCZ/gatsby-transformer-gitinfo",
-    "@loadable/component": "^5.15.2",
+    "@loadable/component": "^5.16.4",
     "adm-zip": "^0.5.10",
     "babel-preset-gatsby": "^3.9.0",
     "bootstrap-italia": "^2.13.2",
